+++ src/components/beta/gux-anchor/example.html
@@ -1,0 +1,50 @@
+<h1>gux-anchor-beta</h1>
+
+<h2>Default</h2>
+<gux-anchor-beta>
+  <a href="#">Link</a>
+</gux-anchor-beta>
+
+<h2>Target</h2>
+<gux-anchor-beta>
+  <a href="#">Open in current window</a>
+</gux-anchor-beta>
+
+<gux-anchor-beta>
+  <a href="#" target="_blank">Open in new tab</a>
+</gux-anchor-beta>
+
+<h2>External Link (Icon)</h2>
+<gux-anchor-beta>
+  <a href="https://spark.genesys.com/" target="_blank">External Link - Spark</a>
+</gux-anchor-beta>
+
+<h2>In Table Cell Link</h2>
+<gux-table-beta>
+  <table slot="data">
+    <thead>
+      <tr>
+        <th data-column-name="first-name">First and Last Name</th>
+        <th data-column-name="last-name">Other</th>
+      </tr>
+    </thead>
+    <tbody>
+      <tr>
+        <td>John Doe</td>
+        <td>
+          <gux-anchor-beta table="true">
+            <a href="#">John Doe Information</a>
+          </gux-anchor-beta>
+        </td>
+      </tr>
+      <tr>
+        <td>Jane Doe</td>
+        <td>
+          <gux-anchor-beta table="true">
+            <a href="#">Jane Doe Information</a>
+          </gux-anchor-beta>
+        </td>
+      </tr>
+    </tbody>
+  </table>
+</gux-table-beta>
