+++ src/components/beta/gux-anchor/gux-anchor.less
@@ -1,0 +1,44 @@
+@import (reference) '../../../style/color.less';
+@import (reference) '../../../style/mixins.less';
+@import (reference) '../../../style/typography.less';
+
+:host {
+  display: flex;
+  align-items: center;
+}
+
+::slotted(a) {
+  color: @gux-blue-60;
+  .gux-anchor();
+}
+
+::slotted(a:hover) {
+  color: @gux-blue-50;
+}
+
+::slotted(a:active) {
+  color: @gux-blue-40;
+}
+
+::slotted(a:focus) {
+  .focus-ring();
+
+  border-radius: 2px;
+}
+
+:host([table]) {
+  ::slotted(a) {
+    .gux-anchor-table();
+  }
+
+  ::slotted(a:hover),
+  ::slotted(a:active),
+  ::slotted(a:focus) {
+    .gux-anchor();
+  }
+}
+
+gux-icon {
+  padding-right: 8px;
+  color: @gux-blue-60;
+}
