+++ src/components/beta/gux-anchor/readme.md
@@ -1,0 +1,30 @@
+# gux-link
+
+
+
+<!-- Auto Generated Below -->
+
+
+## Properties
+
+| Property | Attribute | Description                                            | Type      | Default |
+| -------- | --------- | ------------------------------------------------------ | --------- | ------- |
+| `table`  | `table`   | True when anchor component is used within a table cell | `boolean` | `false` |
+
+
+## Dependencies
+
+### Depends on
+
+- [gux-icon](../../stable/gux-icon)
+
+### Graph
+```mermaid
+graph TD;
+  gux-anchor-beta --> gux-icon
+  style gux-anchor-beta fill:#f9f,stroke:#333,stroke-width:4px
+```
+
+----------------------------------------------
+
+*Built with [StencilJS](https://stenciljs.com/)*
