--- src/components/beta/gux-link/gux-link.less
@@ -1,52 +1,0 @@
-@import (reference) '../../../style/color.less';
-@import (reference) '../../../style/mixins.less';
-
-.gux-link-container {
-  display: flex;
-  align-items: center;
-  font-family: inherit;
-  font-size: 12px;
-  line-height: 1.6667;
-
-  .gux-link {
-    font-family: inherit;
-    font-size: 12px;
-    color: @gux-blue-60;
-    text-decoration: underline;
-
-    &:hover {
-      color: @gux-blue-50;
-    }
-
-    &:active {
-      color: @gux-blue-40;
-    }
-
-    &:focus {
-      .focus-ring();
-
-      border-radius: 2px;
-    }
-
-    &.gux-link-table-cell {
-      text-decoration: none;
-
-      &:hover {
-        text-decoration: underline;
-      }
-
-      &:active {
-        text-decoration: underline;
-      }
-
-      &:focus {
-        text-decoration: underline;
-      }
-    }
-  }
-
-  gux-icon {
-    padding-right: 8px;
-    color: @gux-blue-60;
-  }
-}
