--- src/components/beta/gux-link/gux-link.types.ts
@@ -1,1 +1,0 @@
-export type GuxLinkTargetTypes = '_self' | '_blank' | '_parent' | '_top';
