+++ src/style/style.less
@@ -8,15 +8,6 @@
 @import '../components/beta/gux-form-field-v2/gux-form-field.light.less';
 @import '../components/beta/gux-table/gux-table.light.less';
 
-a {
-  color: @gux-blue-60;
-  text-decoration: none;
-}
-
-a:hover {
-  text-decoration: underline;
-}
-
 gux-action-toast {
   dl {
     dt {
