+++ src/style/typography.less
@@ -102,6 +102,14 @@ h6,
   }
 }
 
+.gux-anchor {
+  text-decoration: underline;
+}
+
+.gux-anchor-table {
+  text-decoration: none;
+}
+
 /* Helpers */
 
 .base-font() {
